requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymupdf==1.24.7
sentence-transformers==3.0.1
faiss-cpu==1.8.0
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy==1.26.4
